requirements.txt
absl-py==2.4.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.128.7
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
gradio==6.5.1
gradio_client==2.0.3
groovy==0.1.2
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
keras==3.12.1
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tomlkit==0.13.3
tqdm==4.67.3
typer==0.21.2
typer-slim==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
wrapt==2.1.1
